twin

+++ Projectile.java
@@ -8,7 +8,7 @@ import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
  */
 
 public class Projectile extends animatedObject {
-    public double speed = 4;
+    public int speed = 4;
     public GreenfootImage[] image;
     public boolean hit = false;
     public MyWorld MyWorld;
@@ -21,7 +21,7 @@ public class Projectile extends animatedObject {
      * Act - do whatever the Pea wants to do. This method is called whenever
      * the 'Act' or 'Run' button gets pressed in the environment.
      */
-    public Projectile(String name, int frameCount, int yPos, int dmg, double speed) {
+    public Projectile(String name, int frameCount, int yPos, int dmg, int speed) {
         this.frameCount = frameCount;
         this.image = importSprites(name, frameCount);
         this.yPos = yPos;
